+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## Unreleased
 
-
+- Add `multi` to Output Objects, as a way to define more than one field of the same type at once.